+++ lib/APIEndpoints.php
@@ -27,21 +27,41 @@ class APIEndpoints {
      * @return array
      *
      * @throws \Wireframe\APIException if no component name was specified (HTTP 400).
+     * @throws \Wireframe\APIException if path has too many parts in it (HTTP 400).
      * @throws \Wireframe\APIException if unknown component was requested (HTTP 404).
      * @throws \Wireframe\APIException if an error occurred while processing the component (HTTP 500).
      */
     public function components(array $path, array $args = []): array {
-        if (empty($path)) {
+
+        // bail out early if path is empty or has too many parts in it
+        if (empty($path) || empty($path[0])) {
             throw (new \Wireframe\APIException('Missing component name'))
                 ->setResponseCode(400);
         }
-        $component_name = $path[0] ?? null;
+        if (isset($path[2])) {
+            throw (new \Wireframe\APIException('Path has too many parts'))
+                ->setResponseCode(400);
+        }
+
+        // get component name from path
+        $component_name = $path[0];
+
+        // set return format
+        $return_format = $path[1] ?? null;
+        if (!in_array($return_format, ['json', 'rendered'])) {
+            $return_format = null;
+        }
+
         try {
             $component = \Wireframe\Factory::component($component_name, $args);
-            return [
-                'json' => json_decode($component->renderJSON()),
-                'rendered' => $component->render(),
-            ];
+            $out = [];
+            if (\is_null($return_format) || $return_format == 'json') {
+                $out['json'] = json_decode($component->renderJSON());
+            }
+            if (\is_null($return_format) || $return_format == 'rendered') {
+                $out['rendered'] = $component->render();
+            }
+            return $out;
         } catch (\Exception $e) {
             if ($e->getCode() === 404) {
                 throw (new \Wireframe\APIException(sprintf(
@@ -59,27 +79,115 @@ class APIEndpoints {
     /**
      * Pages endpoint
      *
-     * Note: this method is a temporary placeholder.
-     *
      * @param array $path
      * @param array $args
      * @return array
+     *
+     * @throws \Wireframe\APIException if no page ID was specified (HTTP 400).
+     * @throws \Wireframe\APIException if path has too many parts in it (HTTP 400).
+     * @throws \Wireframe\APIException if invalid ID was specified (HTTP 400).
+     * @throws \Wireframe\APIException if unknown component was requested (HTTP 404).
+     * @throws \Wireframe\APIException if an error occurred while processing the component (HTTP 500).
      */
     public function pages(array $path, array $args = []): array {
-        return [];
+
+        // bail out early if path is empty or has too many parts in it
+        if (empty($path) || empty($path[0])) {
+            throw (new \Wireframe\APIException('Missing page ID'))
+                ->setResponseCode(400);
+        }
+        if (isset($path[2])) {
+            throw (new \Wireframe\APIException('Path has too many parts'))
+                ->setResponseCode(400);
+        }
+
+        // get page ID from path
+        $page_id = (int) $path[0];
+        if ($page_id === 0) {
+            throw (new \Wireframe\APIException(sprintf(
+                'Invalid page ID (%s)',
+                $path[0]
+            )))->setResponseCode(400);
+        }
+
+        // set return format
+        $return_format = $path[1] ?? null;
+        if (!in_array($return_format, ['json', 'rendered'])) {
+            $return_format = null;
+        }
+
+        try {
+            $page = \Wireframe\Factory::page($page_id, $args);
+            if ($page instanceof \ProcessWire\NullPage || !$page->viewable()) {
+                throw new \ProcessWire\Wire404Exception();
+            }
+            $page->wire('page', $page);
+            $urlSegmentCount = count($page->wire('input')->urlSegments());
+            if ($urlSegmentCount) {
+                for ($urlSegmentNum = 0; $urlSegmentNum < $urlSegmentCount; $urlSegmentNum++) {
+                    // note: num is always 1 since removing the first segment resets keys for the rest of them.
+                    $page->wire('input')->setUrlSegment(1, null);
+                }
+            }
+            $out = [];
+            if (\is_null($return_format) || $return_format == 'json') {
+                $controller = $page->getController();
+                $out['json'] = $controller ? json_decode($controller->renderJSON()) : $page->getArray();
+            }
+            if (\is_null($return_format) || $return_format == 'rendered') {
+                $out['rendered'] = $page->render();
+            }
+            return $out;
+        } catch (\Exception $e) {
+            if ($e instanceof \ProcessWire\Wire404Exception) {
+                throw (new \Wireframe\APIException(sprintf(
+                    'Unknown page (id=%s)',
+                    $page_id
+                )))->setResponseCode(404);
+            }
+            throw (new \Wireframe\APIException(sprintf(
+                'Error while processing page (id=%s)',
+                $page_id
+            )));
+        }
     }
 
     /**
      * Partials endpoint
      *
-     * Note: this method is a temporary placeholder.
-     *
      * @param array $path
      * @param array $args
      * @return array
+     *
+     * @throws \Wireframe\APIException if no partial name was specified (HTTP 400).
      */
     public function partials(array $path, array $args = []): array {
-        return [];
+
+        // bail out early if path is empty or has too many parts in it
+        if (empty($path) || empty($path[0])) {
+            throw (new \Wireframe\APIException('Missing partial name'))
+                ->setResponseCode(400);
+        }
+
+        // get partial name from path
+        $partial_name = implode('/', $path);
+
+        try {
+            return [
+                'rendered' => \Wireframe\Factory::partial($partial_name, $args),
+            ];
+        } catch (\Exception $e) {
+            if ($e->getCode() === 404) {
+                throw (new \Wireframe\APIException(sprintf(
+                    'Unknown partial (%s)',
+                    $e->getMessage()
+                )))->setResponseCode(404);
+            }
+            throw (new \Wireframe\APIException(sprintf(
+                'Error while processing partial (%s)',
+                $partial_name
+            )));
+        }
     }
 
 }